server/requirements.txt
fastapi[standard]
uvicorn
pinecone
pypdf
langchain
google-genai 